Binomial counts for Bayesian Adaptive Trials

Simulation for binomial counts for Bayesian Adaptive trial with different inputs to control for power, sample size, type 1 error rate, etc.

binomialBACT(p_treatment, p_control = NULL, y0_treatment = NULL,
  N0_treatment = NULL, y0_control = NULL, N0_control = NULL,
  discount_function = "identity", N_total, lambda, lambda_time,
  interim_look = NULL, EndofStudy, prior = c(1, 1), block = 2,
  rand_ratio = c(1, 1), prop_loss_to_followup = 0.1,
  alternative = "two-sided", h0 = 0, futility_prob = 0.05,
  expected_success_prob = 0.9, prob_ha = 0.95, N_impute = 1000,
  number_mcmc = 1000)

Arguments

p_treatment
scalar. Proportion of events under the treatment arm.

p_control
scalar. Proportion of events under the control arm.

y0_treatment
scalar. Number of events for the historical treatment arm.

N0_treatment
scalar. Sample size of the historical treatment arm.

y0_control
scalar. Number of events for the historical control arm.

N0_control
scalar. Sample size of the historical control arm.

discount_function
character. Specify the discount function to use. Currently supports the Weibull function (discount_function="weibull"), the scaled-Weibull function (discount_function="scaledweibull"), and the identity function (discount_function="identity"). The scaled-Weibull discount function scales the output of the Weibull CDF to have a max value of 1. The identity discount function uses the posterior probability directly as the discount weight. Default value is "identity".

N_total
scalar. Total sample size.

lambda
vector. Lambda for different enrollment rates across times.

lambda_time
vector. Same size as lambda, denote times at lambda changes.

interim_look
vector. Sample size for interim looks. Note: the overall trial size should not be included.

EndofStudy
scalar. Length of the study.

prior
vector. Prior value of beta rate, beta(a0, b0).

block
scalar. Block size for randomization to be implemented.

rand_ratio
vector. Randomization allocation for control to treatment. Integer values mapping the size of the block.

prop_loss_to_followup
scalar. Proportion of subjects lost to follow-up.

alternative
character. The string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".

h0
scalar. Treshold for comparing two proportions. Default is h0=0.

futility_prob
scalar. Type I error rate.

expected_success_prob
scalar. Power of the study.

prob_ha
scalar. Probability of alternative hypothesis.

N_impute
scalar. Number of imputations for Monte Carlo simulation of missing data.

number_mcmc
scalar. Number of Monte Carlo Markov Chain draws in sampling posterior.

Value

a list of output

Examples

binomialBACT(p_control = 0.12, p_treatment = 0.10,
             y0_treatment = 8, N0_treatment = 90,
             y0_control = 13, N0_control = 95,
             N_total = 300, N_impute = 100,
             lambda = c(0.3, 1), lambda_time = c(25),
             interim_look = c(210, 240, 270),
             EndofStudy = 50)

#> [1] 210
#> [1] 240
#> [1] 270
#> [1] 300

#> $p_treatment
#> [1] 0.1
#> 
#> $p_control
#> [1] 0.12
#> 
#> $prob_of_accepting_alternative
#> [1] 0.95
#> 
#> $margin
#> [1] 0
#> 
#> $interim_look
#> [1] 210 240 270
#> 
#> $N_treatment
#> [1] 138
#> 
#> $N_control
#> [1] 132
#> 
#> $N_complete
#> [1] 270
#> 
#> $N_enrolled
#> [1] 300
#> 
#> $N_max
#> [1] 300
#> 
#> $post_prob_accept_alternative
#> [1] 0.946
#> 
#> $est_final
#> [1] -0.05560319
#> 
#> $est_interim
#> [1] -0.03217806
#> 
#> $stop_futility
#> [1] 0
#> 
#> $stop_expected_success
#> [1] 0
#>