Data file for normal analysis

Wrapper function for data file in normal analysis.

data_normal(data, .data = NULL)

Arguments

data
data frame. A data frame which provides patient id, treatment group, outcome of the treatment and complete columns. An example file is available at data(normaldata).

.data
NULL. stores the normal data for analysis, please do not fill it in.

Value

a list with normal data file to be analyzed.

Examples

data_normal(data = normaldata)

#> $data
#>             id treatment    outcome complete
#> 1     Patient1         1 13.7021906        1
#> 2     Patient2         1 11.9255183        1
#> 3     Patient3         1 10.5869434        1
#> 4     Patient4         1 10.4500312        1
#> 5     Patient5         1  7.1546154        1
#> 6     Patient6         1  8.8743957        0
#> 7     Patient7         1  8.8118973        1
#> 8     Patient8         0 10.3945772        1
#> 9     Patient9         1  7.4870376        1
#> 10   Patient10         1  7.4485120        1
#> 11   Patient11         1 12.1894188        1
#> 12   Patient12         0 12.1382648        1
#> 13   Patient13         1  9.9979052        1
#> 14   Patient14         1  9.9385510        1
#> 15   Patient15         1  9.6532271        1
#> 16   Patient16         0 11.8667983        1
#> 17   Patient17         0 14.1727265        1
#> 18   Patient18         1  9.6205222        0
#> 19   Patient19         1 11.1914697        1
#> 20   Patient20         1 11.8370766        1
#> 21   Patient21         1 14.7301169        1
#> 22   Patient22         0 13.4419500        1
#> 23   Patient23         1  9.5864860        1
#> 24   Patient24         1  8.9138521        1
#> 25   Patient25         1  6.8576292        1
#> 26   Patient26         0  9.2750803        1
#> 27   Patient27         0 11.4873621        1
#> 28   Patient28         1 10.3558705        1
#> 29   Patient29         0 11.8202970        0
#> 30   Patient30         0 15.7474901        1
#> 31   Patient31         0 12.4498969        1
#> 32   Patient32         1  9.4496192        1
#> 33   Patient33         1 11.4885826        1
#> 34   Patient34         1  9.3555919        1
#> 35   Patient35         1  9.6573159        1
#> 36   Patient36         0 12.3585446        1
#> 37   Patient37         1 11.4993190        1
#> 38   Patient38         0 12.4479771        1
#> 39   Patient39         1  7.1827550        1
#> 40   Patient40         0 12.6452215        1
#> 41   Patient41         1 10.8559874        1
#> 42   Patient42         1  8.1421044        0
#> 43   Patient43         0  7.1019621        1
#> 44   Patient44         0 15.0428834        1
#> 45   Patient45         0 13.1871503        1
#> 46   Patient46         0 10.1664984        0
#> 47   Patient47         1  8.7606257        1
#> 48   Patient48         1  7.2938045        1
#> 49   Patient49         1 11.1202104        1
#> 50   Patient50         1 11.3533487        1
#> 51   Patient51         1 12.9789731        1
#> 52   Patient52         0 14.7882798        1
#> 53   Patient53         1  7.0401467        1
#> 54   Patient54         0 12.5937207        1
#> 55   Patient55         1  9.6169556        1
#> 56   Patient56         1 12.3220162        1
#> 57   Patient57         0 13.9689015        1
#> 58   Patient58         1 10.2625238        1
#> 59   Patient59         1  9.9879829        1
#> 60   Patient60         1 12.0305052        1
#> 61   Patient61         0 15.5897618        1
#> 62   Patient62         0 11.3535626        1
#> 63   Patient63         1 12.7412761        1
#> 64   Patient64         1  9.7779481        1
#> 65   Patient65         0 11.7016925        1
#> 66   Patient66         0 13.9377849        1
#> 67   Patient67         0 13.2373351        1
#> 68   Patient68         0 12.1895451        1
#> 69   Patient69         1 12.6508586        0
#> 70   Patient70         1 13.2062960        0
#> 71   Patient71         0 14.1220440        1
#> 72   Patient72         0 14.8487508        1
#> 73   Patient73         1 13.0078694        1
#> 74   Patient74         0 13.9715172        0
#> 75   Patient75         0  7.2501986        1
#> 76   Patient76         1 10.0236651        1
#> 77   Patient77         1  5.8327090        0
#> 78   Patient78         0 11.9337668        1
#> 79   Patient79         1  9.2379120        1
#> 80   Patient80         1 10.7514631        1
#> 81   Patient81         0 15.4649258        1
#> 82   Patient82         1  7.3017105        0
#> 83   Patient83         1  9.8781634        1
#> 84   Patient84         0  9.2690927        1
#> 85   Patient85         0 11.1862125        1
#> 86   Patient86         0 14.1435845        1
#> 87   Patient87         1  9.6411837        1
#> 88   Patient88         0 12.5678831        1
#> 89   Patient89         1  9.6927252        1
#> 90   Patient90         0 13.0493123        0
#> 91   Patient91         0  8.0583374        0
#> 92   Patient92         1  9.2074806        1
#> 93   Patient93         0 13.4992637        1
#> 94   Patient94         1 10.3442688        1
#> 95   Patient95         1 13.1037878        1
#> 96   Patient96         0 14.0204119        1
#> 97   Patient97         0 13.3424295        1
#> 98   Patient98         1 10.1460779        1
#> 99   Patient99         1  9.9131479        1
#> 100 Patient100         0 11.6273345        1
#> 101 Patient101         0 13.4867567        1
#> 102 Patient102         0 11.8521467        1
#> 103 Patient103         1 13.3924460        1
#> 104 Patient104         0 11.8932416        0
#> 105 Patient105         0 12.9983557        1
#> 106 Patient106         0 16.5311440        1
#> 107 Patient107         0 12.4883828        1
#> 108 Patient108         1  5.7133837        1
#> 109 Patient109         1 12.6170609        1
#> 110 Patient110         1  8.1151905        1
#> 111 Patient111         0 12.9654652        1
#> 112 Patient112         1  6.1250571        1
#> 113 Patient113         0 13.5360220        1
#> 114 Patient114         0 13.3173224        0
#> 115 Patient115         1  9.6419254        1
#> 116 Patient116         0 12.9592200        1
#> 117 Patient117         0 15.7663572        0
#> 118 Patient118         1 10.2025243        1
#> 119 Patient119         1 11.5544171        1
#> 120 Patient120         1  9.2202067        1
#> 121 Patient121         0 13.3960809        1
#> 122 Patient122         1  7.8477475        1
#> 123 Patient123         0 11.0716213        1
#> 124 Patient124         1 10.3981986        1
#> 125 Patient125         1  9.9173010        1
#> 126 Patient126         0 13.9147880        1
#> 127 Patient127         1  6.1111475        1
#> 128 Patient128         0 17.1003049        1
#> 129 Patient129         0 13.9079879        1
#> 130 Patient130         1  7.3362985        1
#> 131 Patient131         0 14.5493762        0
#> 132 Patient132         0 16.2000770        1
#> 133 Patient133         1  9.6000736        1
#> 134 Patient134         1  5.6003980        1
#> 135 Patient135         0 10.6623887        0
#> 136 Patient136         1 12.8959507        0
#> 137 Patient137         1  9.2863935        1
#> 138 Patient138         0 14.3840639        1
#> 139 Patient139         1  9.8862610        1
#> 140 Patient140         0 11.4884991        1
#> 141 Patient141         0 12.8195856        1
#> 142 Patient142         0 12.3898229        1
#> 143 Patient143         0 14.0686891        1
#> 144 Patient144         1 12.4019542        1
#> 145 Patient145         1  8.9183182        1
#> 146 Patient146         0 12.9698047        1
#> 147 Patient147         0 12.5461852        1
#> 148 Patient148         0 12.3979965        1
#> 149 Patient149         1  8.6574664        1
#> 150 Patient150         0 12.8481174        1
#> 151 Patient151         0 14.4524260        1
#> 152 Patient152         0 11.9761602        1
#> 153 Patient153         1 11.3627882        1
#> 154 Patient154         0 11.9094208        0
#> 155 Patient155         1 10.2056626        1
#> 156 Patient156         0 12.7035719        1
#> 157 Patient157         1  7.2154949        1
#> 158 Patient158         1 10.1687268        1
#> 159 Patient159         0 11.3219147        0
#> 160 Patient160         1 11.5110468        0
#> 161 Patient161         1  9.3418099        1
#> 162 Patient162         0 14.7820743        1
#> 163 Patient163         1  9.7927162        1
#> 164 Patient164         0  8.7650016        0
#> 165 Patient165         1  6.5398416        1
#> 166 Patient166         1 10.2822322        1
#> 167 Patient167         1 11.2542806        1
#> 168 Patient168         1 12.4857294        1
#> 169 Patient169         0 12.9174595        1
#> 170 Patient170         1 12.4278293        1
#> 171 Patient171         0 12.5008056        1
#> 172 Patient172         0 15.0813265        0
#> 173 Patient173         0 14.8896322        1
#> 174 Patient174         1  8.1423906        1
#> 175 Patient175         1 13.0721731        1
#> 176 Patient176         1  7.2973438        1
#> 177 Patient177         0 17.8219652        1
#> 178 Patient178         0 17.6658299        1
#> 179 Patient179         0 10.2374539        1
#> 180 Patient180         0 10.3481287        1
#> 181 Patient181         0 11.6406321        1
#> 182 Patient182         1  9.9194506        1
#> 183 Patient183         0  9.9953076        1
#> 184 Patient184         0 12.3320632        1
#> 185 Patient185         1 10.7022459        1
#> 186 Patient186         0  8.9105274        1
#> 187 Patient187         1 10.6311936        1
#> 188 Patient188         0 13.4741000        1
#> 189 Patient189         0  9.1315673        1
#> 190 Patient190         1  9.3549431        0
#> 191 Patient191         1 10.4791168        1
#> 192 Patient192         0 15.8223202        0
#> 193 Patient193         0 10.1548152        1
#> 194 Patient194         0 13.6121257        1
#> 195 Patient195         0 13.4306095        1
#> 196 Patient196         0 14.4217976        0
#> 197 Patient197         1  8.5416284        1
#> 198 Patient198         1 14.5642550        1
#> 199 Patient199         1  6.4848850        1
#> 200 Patient200         0 13.0737234        1
#> 201 Patient201         0 11.9238560        0
#> 202 Patient202         1  9.7271310        1
#> 203 Patient203         1  8.4406959        1
#> 204 Patient204         0 13.0279008        1
#> 205 Patient205         0 13.3262545        1
#> 206 Patient206         1 10.4640729        1
#> 207 Patient207         1  6.2087697        1
#> 208 Patient208         0 11.8745366        1
#> 209 Patient209         0 12.1453277        1
#> 210 Patient210         0 14.3170134        1
#> 211 Patient211         1 10.8991490        1
#> 212 Patient212         0 11.3452388        1
#> 213 Patient213         1 11.7556543        1
#> 214 Patient214         0 11.7239612        1
#> 215 Patient215         0 13.3334806        0
#> 216 Patient216         1  6.6803236        1
#> 217 Patient217         0 13.8320414        1
#> 218 Patient218         0 15.7716533        1
#> 219 Patient219         1  5.7475308        1
#> 220 Patient220         0  9.8415097        1
#> 221 Patient221         1 11.2258347        1
#> 222 Patient222         1 11.3210413        1
#> 223 Patient223         0 16.6037331        1
#> 224 Patient224         1  8.9484827        1
#> 225 Patient225         1 10.9433568        1
#> 226 Patient226         0 15.0488201        1
#> 227 Patient227         1  8.3138094        1
#> 228 Patient228         1 11.4340631        1
#> 229 Patient229         1  7.8630281        0
#> 230 Patient230         1 10.0150064        1
#> 231 Patient231         0 12.8919722        1
#> 232 Patient232         0 14.2839906        1
#> 233 Patient233         1 11.1802820        1
#> 234 Patient234         1  0.6751561        1
#> 235 Patient235         1 11.6933759        0
#> 236 Patient236         0 14.1722926        1
#> 237 Patient237         0 13.1867031        1
#> 238 Patient238         1  9.8939632        1
#> 239 Patient239         0 13.8321494        1
#> 240 Patient240         1 10.7100894        0
#> 241 Patient241         1 10.8251780        1
#> 242 Patient242         0 10.8932647        1
#> 243 Patient243         1  9.7978750        1
#> 244 Patient244         0 12.3983009        0
#> 245 Patient245         0 12.2834200        1
#> 246 Patient246         0 11.7901332        1
#> 247 Patient247         0 14.7665796        1
#> 248 Patient248         1  9.9950031        1
#> 249 Patient249         1  9.3424064        1
#> 250 Patient250         0 11.6889777        1
#> 251 Patient251         1 11.7138390        1
#> 252 Patient252         0 12.0071214        1
#> 253 Patient253         0 14.1452696        1
#> 254 Patient254         0 14.5510806        1
#> 255 Patient255         1  9.6420809        1
#> 256 Patient256         1  8.2636419        1
#> 257 Patient257         0 13.2769877        1
#> 258 Patient258         1  6.1998823        1
#> 259 Patient259         1 11.4270568        1
#> 260 Patient260         1 11.5179158        1
#> 261 Patient261         1 11.3097331        1
#> 262 Patient262         1  8.3638321        1
#> 263 Patient263         1  7.3501157        1
#> 264 Patient264         0 13.6014666        1
#> 265 Patient265         0 10.8065193        1
#> 266 Patient266         1 10.3922310        1
#> 267 Patient267         0 15.3717355        1
#> 268 Patient268         0 11.6590808        1
#> 269 Patient269         1 10.1423512        1
#> 270 Patient270         0 11.7790321        1
#> 271 Patient271         0 15.0630558        1
#> 272 Patient272         0 12.5011350        1
#> 273 Patient273         0 14.3734310        1
#> 274 Patient274         0 12.5833765        1
#> 275 Patient275         0 16.1172286        1
#> 276 Patient276         0 13.8548295        1
#> 277 Patient277         1 10.8768534        1
#> 278 Patient278         0 11.0936980        1
#> 279 Patient279         1 11.8892349        1
#> 280 Patient280         0 12.6034363        1
#> 281 Patient281         0 10.4276531        1
#> 282 Patient282         1  8.7033430        1
#> 283 Patient283         0 11.4000031        1
#> 284 Patient284         0 11.2931605        1
#> 285 Patient285         1 12.8449117        1
#> 286 Patient286         1 11.6445108        1
#> 287 Patient287         1 12.7447186        1
#> 288 Patient288         1  5.4706564        1
#> 289 Patient289         0 11.7828728        0
#> 290 Patient290         0 11.7962795        1
#> 291 Patient291         1  6.6183661        1
#> 292 Patient292         1  8.6650788        1
#> 293 Patient293         0 14.6637751        1
#> 294 Patient294         1 11.3878399        1
#> 295 Patient295         0 11.5924309        1
#> 296 Patient296         1 11.0639694        1
#> 297 Patient297         1  8.4297614        1
#> 298 Patient298         1  9.6749568        1
#> 299 Patient299         0 12.8678958        1
#> 300 Patient300         0 12.5586880        0
#>