Simulating enrollment dates

This function simulates enrollment dates using either poisson distribution

enrollment(param, N_total, time = NULL)

Arguments

param
a vector of lambda in poisson

N_total
a numeric value of total sample size

time
a vector of the length(param) - 1 indicating end of time when a specific lambda is used

Value

a vector of enrollment dates

Examples

enrollment(param = c(0.003, 0.7), 100, time = 10)

#>   [1]   0   0   1   3   5   6   7   9  11  11  12  13  14  16  18  28  29  29
#>  [19]  30  31  31  32  35  36  37  39  40  41  43  43  43  44  45  46  46  47
#>  [37]  48  50  54  55  56  57  57  60  62  62  63  64  65  66  67  69  71  75
#>  [55]  77  78  80  81  89  89  89  91  91  97  98  98 102 102 102 104 105 109
#>  [73] 110 110 110 110 111 111 112 115 117 117 118 121 123 123 125 126 127 129
#>  [91] 130 134 138 138 139 139 140 141 142 143

enrollment(param = c(0.3, 0.5, 0.9, 1.2, 2.1), 200, c(20, 30, 40, 60))

#>   [1]   0   4   4   5   5   6  13  13  15  16  16  16  16  17  18  18  20  24
#>  [19]  25  25  25  26  33  34  36  36  37  37  38  40  42  42  44  44  46  47
#>  [37]  47  48  52  52  52  52  53  53  55  56  57  57  58  58  58  58  58  59
#>  [55]  59  60  61  61  61  61  62  62  63  63  63  64  65  67  67  67  68  68
#>  [73]  68  69  69  69  71  72  73  73  75  76  76  76  76  78  79  79  79  80
#>  [91]  80  80  80  81  81  82  82  82  82  82  84  84  84  84  84  85  85  85
#> [109]  85  85  86  87  87  87  88  88  89  89  89  89  89  90  90  90  90  91
#> [127]  91  92  92  94  95  95  95  95  96  96  97  97  97  98 100 100 100 102
#> [145] 102 102 103 103 103 103 104 104 104 105 105 106 107 107 107 108 108 109
#> [163] 109 109 109 109 110 110 111 111 111 112 113 113 114 114 115 116 116 117
#> [181] 117 117 118 119 119 119 119 120 121 122 123 123 124 124 124 125 125 125
#> [199] 126 127