Historical data for normal distribution

Wrapper function for historical data from normal outcome.

historical_normal(mu0_treatment = NULL, sd0_treatment = NULL,
  N0_treatment = NULL, mu0_control = NULL, sd0_control = NULL,
  N0_control = NULL, discount_function = "identity", .data = NULL)

Arguments

mu0_treatment
numeric. Mean of the historical treatment group.

sd0_treatment
numeric. The Standard deviation of the historical treatment group.

N0_treatment
numeric. scalar. Number of observations of the historical treatment group.

mu0_control
numeric. Mean of the historical control group.

sd0_control
numeric. The Standard deviation of the historical control group.

N0_control
numeric. umber of observations of the historical control group.

discount_function
character. Specify the discount function to use. Currently supports weibull, scaledweibull, and identity. The discount function scaledweibull scales the output of the Weibull CDF to have a max value of 1. The identity discount function uses the posterior probability directly as the discount weight. Default value is "identity".

.data
NULL. stores the proportion of control and treatment, please do not fill it in.

Value

a list with historical data for control and treatment group with the discount function.

Examples

historical_normal(mu0_treatment = 15, sd0_treatment = 2, N0_treatment = 10,
                            mu0_control = 17, sd0_control = 3, N0_control = 20)

#> $mu0_treatment
#> [1] 15
#> 
#> $sd0_treatment
#> [1] 2
#> 
#> $N0_treatment
#> [1] 10
#> 
#> $mu0_control
#> [1] 17
#> 
#> $sd0_control
#> [1] 3
#> 
#> $N0_control
#> [1] 20
#> 
#> $discount_function
#> [1] "identity"
#>