Imputes time-to-event outcomes.

Imputation of time-to-event outcomes using the piecewise constant hazard exponential function.

pw_exp_impute(time, hazard, maxtime = NULL, cutpoint = NULL)

Arguments

time
vector. The observed time for patient that have had no event or passed maxtime.

hazard
vector. The constant hazard rates for exponential failures.

maxtime
scalar. maximum time before end of study.

cutpoint
vector. The change-point vector indicating time when the hazard rates change.

Value

a dataset with simulated follow-up time (time) and respective event indicator (1 = event, 0 = censoring)

Examples

pw_exp_impute(time = c(120), c(0.005, 0.001), 110, 40)

#>   time event
#> 1  110     0

          pw_exp_impute(time = c(10, 20, 30), c(0.005, 0.01, 0.02), 100, c(40, 80))

#>        time event
#> 1  16.08357     1
#> 2 100.00000     0
#> 3 100.00000     0

          pw_exp_impute(time = c(40, 30), c(0.005, 0.01), 120, c(50))

#>       time event
#> 1  67.8358     1
#> 2 120.0000     0