Analyzing Bayesian trial for time-to-event data

Function to analyze Bayesian trial for time-to-event data which allows early stopping and incorporation of historical data using the discount function approach

survival_analysis(time, treatment, event = NULL, time0 = NULL,
  treatment0 = NULL, event0 = NULL, surv_time = NULL, h0 = 0,
  breaks = NULL, alternative = "greater", N_impute = 10,
  number_mcmc = 10000, prob_ha = 0.95, futility_prob = 0.1,
  expected_success_prob = 0.9, prior = c(0.1, 0.1),
  discount_function = "identity", fix_alpha = FALSE, alpha_max = 1,
  weibull_scale = 0.135, weibull_shape = 3, method = "fixed")

Arguments

time
vector. exposure time for the subjects. It must be the same length as the treatment variable.

treatment
vector. treatment assignment for patients, 1 for treatment group and 0 for control group

event
vector. The status indicator, normally 0=alive, 1=dead. Other choices are TRUE/FALSE (TRUE = death) or 1/2 (2=death). For censored data, the status indicator is 0=right censored, 1 = event at time. Although unusual, the event indicator can be omitted, in which case all subjects are assumed to have an event.

time0
vector. Historical exposure time for the subjects. It must be the same length as the treatment variable.

treatment0
vector. the historical treatment assignment for patients, 1 for treatment group and 0 for control group.

event0
vector. Historical status indicator, normally 0=alive, 1=dead. Other choices are TRUE/FALSE (TRUE = death) or 1/2 (2=death). For censored data, the status indicator is 0=right censored, 1 = event at time. Although unusual, the event indicator can be omitted, in which case all subjects are assumed to have an event.

surv_time
scalar. scalar. Survival time of interest for computing the probability of survival for a single arm (OPC) trial. Default is overall, i.e., current+historical, median survival time.

h0
scalar. Threshold for comparing two mean values. Default is h0=0.

breaks
vector. Breaks (interval starts) used to compose the breaks of the piecewise exponential model. Do not include zero. Default breaks are the quantiles of the input times.

alternative
character. The string specifying the alternative hypothesis, must be one of "greater" (default), "less" or "two.sided".

N_impute
scalar. Number of imputations for Monte Carlo simulation of missing data.

number_mcmc
scalar. Number of Monte Carlo Markov Chain draws in sampling posterior.

prob_ha
scalar. Probability of alternative hypothesis.

futility_prob
scalar. Probability of stopping early for futility.

expected_success_prob
scalar. Probability of stopping early for success.

prior
vector. Prior values of the gamma rate, Gamma(a0, b0). The default is set to Gamma(.1, .1).

discount_function
character. If incorporating historical data, specify the discount function. Currently supports the Weibull function (discount_function="weibull"), the scaled-Weibull function (discount_function="scaledweibull"), and the identity function (discount_function="identity"). The scaled-Weibull discount function scales the output of the Weibull CDF to have a max value of 1. The identity discount function uses the posterior probability directly as the discount weight. Default value is "identity". See bdpnormal for more details.

fix_alpha
logical. Fix alpha at alpha_max? Default value is FALSE.

alpha_max
scalar. Maximum weight the discount function can apply. Default is 1. For a two-arm trial, users may specify a vector of two values where the first value is used to weight the historical treatment group and the second value is used to weight the historical control group.

weibull_scale
scalar. Scale parameter of the Weibull discount function used to compute alpha, the weight parameter of the historical data. Default value is 0.135. For a two-arm trial, users may specify a vector of two values where the first value is used to estimate the weight of the historical treatment group and the second value is used to estimate the weight of the historical control group. Not used when discount_function = "identity".

weibull_shape
scalar. Shape parameter of the Weibull discount function used to compute alpha, the weight parameter of the historical data. Default value is 3. For a two-arm trial, users may specify a vector of two values where the first value is used to estimate the weight of the historical treatment group and the second value is used to estimate the weight of the historical control group. Not used when discount_function = "identity".

method
character. Analysis method with respect to estimation of the weight paramter alpha. Default method "mc" estimates alpha for each Monte Carlo iteration. Alternate value "fixed" estimates alpha once and holds it fixed throughout the analysis. See the the bdpsurvival vignette
vignette("bdpsurvival-vignette", package="bayesDP") for more details.

Value

a list of output for the Bayesian trial for time-to-event.

prob_of_accepting_alternative

scalar. The input parameter of probability of accepting the alternative.

margin

scalar. The margin input value of difference between mean estimate of treatment and mean estimate of the control.

alternative

character. The input parameter of alternative hypothesis.

alpha_max

scalar. The alpha_max input.

N_treatment

scalar. The number of patients enrolled in the experimental group for each simulation.

event_treatment

scalar. The number of events in the experimental group for each simulation.

N_control

scalar. The number of patients enrolled in the control group for each simulation.

event_control

scalar. The number of events in the control group for each simulation.

N_enrolled

scalar. The number of patients enrolled in the trial (sum of control and experimental group for each simulation. )

N_complete

scalar. The number of patients whose time passes the surv_time.

alpha_discount

vector. The alpha discount funtion used for control and treatment.

post_prob_accept_alternative

vector. The final probability of accepting the alternative hypothesis after the analysis is done.

est_final

scalar. The final estimate of the difference in posterior estimate of treatment and posterior estimate of the control group.

stop_futility

scalar. Did the trial stop for futility during imputation of patient who had loss to follow up? 1 for yes and 0 for no.

stop_expected_success

scalar. Did the trial stop for early success during imputation of patient who had loss to follow up? 1 for yes and 0 for no.