Simulating enrollment dates

This function simulates enrollment dates using either poisson or binomial distribution

enrollment(param, N_total, time = NULL)

Arguments

param
a vector of lambda in poisson

N_total
a numeric value of total sample size

time
a vector of the length(param) - 1 indicating end of time when a specific lambda is used

Value

a vector of enrollment dates

Examples

enrollment(param = c(0.003, 0.7), 100, time = 10)

#>   [1]   0  11  14  14  17  19  20  20  25  31  31  31  33  33  34  37  42  43
#>  [19]  43  43  46  48  49  50  50  51  51  52  52  52  53  54  54  56  56  56
#>  [37]  57  59  61  63  66  69  71  72  73  74  74  75  75  76  77  81  82  85
#>  [55]  85  87  88  89  90  90  93  96  97  99  99  99 100 101 103 106 107 107
#>  [73] 109 112 115 115 115 118 120 120 122 123 124 124 129 134 136 136 137 137
#>  [91] 137 137 138 140 140 142 145 146 148 149

enrollment(param = c(0.3, 0.5, 0.9, 1.2, 2.1), 200, c(20, 30, 40, 60))

#>   [1]   0   1   1  11  14  17  20  28  31  32  32  35  35  37  39  39  39  40
#>  [19]  41  41  45  47  47  47  48  48  49  49  51  51  51  52  52  52  53  54
#>  [37]  54  54  57  57  58  59  59  60  60  60  61  61  61  61  62  62  63  63
#>  [55]  64  64  64  66  68  69  69  69  70  70  70  70  71  71  71  71  71  71
#>  [73]  72  72  72  72  73  74  74  74  75  75  75  76  76  76  77  77  77  78
#>  [91]  78  78  78  78  79  79  79  80  81  81  82  82  82  83  83  83  84  84
#> [109]  84  84  84  84  85  85  87  88  88  88  88  89  89  90  91  91  91  91
#> [127]  92  94  94  94  94  94  95  96  96  96  96  97  98  98  99  99 100 100
#> [145] 101 101 101 105 105 106 106 107 107 107 107 108 108 109 109 110 110 111
#> [163] 111 111 112 113 113 113 114 114 114 114 115 115 117 119 119 119 120 120
#> [181] 121 121 122 122 122 122 123 123 123 123 124 125 125 125 125 126 126 126
#> [199] 127 127