Simulating enrollment dates

This function simulates enrollment dates using either poisson distribution

enrollment(param, N_total, time = NULL)

Arguments

param
a vector of lambda in poisson

N_total
a numeric value of total sample size

time
a vector of the length(param) - 1 indicating end of time when a specific lambda is used

Value

a vector of enrollment dates

Examples

enrollment(param = c(0.003, 0.7), 100, time = 10)

#>   [1]   0   0   2   4   5   6   8   8  11  12  13  15  17  18  20  20  23  23
#>  [19]  25  26  31  32  33  35  43  45  45  48  49  49  51  61  68  69  69  72
#>  [37]  74  74  77  79  79  81  82  86  87  87  88  89  94  96  99 101 104 104
#>  [55] 105 107 108 108 109 110 111 111 112 113 113 115 118 119 119 121 128 129
#>  [73] 130 133 135 137 138 140 141 141 142 143 145 146 146 146 147 147 147 147
#>  [91] 148 153 153 153 154 156 158 158 158 159

enrollment(param = c(0.3, 0.5, 0.9, 1.2, 2.1), 200, c(20, 30, 40, 60))

#>   [1]   0   4  12  14  17  19  20  21  22  24  26  26  26  26  26  28  31  32
#>  [19]  32  32  33  35  36  42  43  44  44  44  44  46  48  49  50  51  52  53
#>  [37]  54  55  56  58  58  58  59  59  59  59  59  59  59  60  61  62  62  63
#>  [55]  63  64  64  64  66  66  66  67  67  67  67  67  69  69  69  70  70  71
#>  [73]  71  71  72  73  73  73  73  73  74  74  74  74  75  75  75  76  76  77
#>  [91]  77  77  78  78  78  79  80  80  80  80  81  81  81  81  82  82  83  83
#> [109]  84  84  85  85  86  86  87  88  88  89  89  89  90  90  91  91  92  92
#> [127]  92  93  94  95  95  95  96  96  96  96  96  97  97  97  97  97  98  98
#> [145] 100 100 100 100 100 101 102 103 103 103 103 103 103 104 104 104 105 105
#> [163] 105 105 106 106 107 107 107 107 108 108 108 109 109 110 110 110 110 111
#> [181] 112 112 113 113 113 114 115 115 115 116 117 117 117 118 118 119 119 119
#> [199] 120 120