Simulating enrollment dates

This function simulates enrollment dates using either poisson distribution

enrollment(param, N_total, time = NULL)

Arguments

param
a vector of lambda in poisson

N_total
a numeric value of total sample size

time
a vector of the length(param) - 1 indicating end of time when a specific lambda is used

Value

a vector of enrollment dates

Examples

enrollment(param = c(0.003, 0.7), 100, time = 10)

#>   [1]   0   1   2   3   5   6   7   8  12  12  14  16  16  18  18  19  22  23
#>  [19]  24  24  25  26  27  28  28  28  28  32  32  32  34  35  35  36  37  37
#>  [37]  37  40  45  48  49  49  51  51  51  52  52  56  56  57  59  62  64  64
#>  [55]  64  66  67  67  68  68  71  72  72  72  75  77  77  77  77  81  81  84
#>  [73]  87  88  88  89  89  90  91  92  95  99 105 110 119 119 122 124 127 130
#>  [91] 132 132 132 133 136 141 146 147 150 151

enrollment(param = c(0.3, 0.5, 0.9, 1.2, 2.1), 200, c(20, 30, 40, 60))

#>   [1]   0   0   4   7   9  12  13  13  15  17  17  22  24  26  30  31  31  34
#>  [19]  35  35  35  36  36  37  42  44  44  44  46  46  47  47  47  50  51  52
#>  [37]  54  54  54  55  55  58  59  60  61  61  61  61  62  62  63  64  64  64
#>  [55]  64  64  65  65  65  65  66  67  67  68  69  69  70  70  70  71  72  72
#>  [73]  72  73  74  74  74  74  75  75  75  75  76  76  76  77  77  78  79  79
#>  [91]  79  80  81  81  81  82  82  83  83  83  84  84  84  84  84  84  85  85
#> [109]  87  87  88  88  89  90  90  91  92  92  92  93  93  93  94  94  94  95
#> [127]  95  95  95  96  97  97  97  97  97  98  98  99  99 100 101 102 103 103
#> [145] 103 103 103 104 105 106 107 107 107 107 108 109 110 110 111 111 112 112
#> [163] 113 113 113 113 114 115 117 117 117 118 118 118 118 119 120 121 121 121
#> [181] 122 122 122 123 123 124 125 125 126 126 126 126 126 126 126 127 127 127
#> [199] 128 129