Simulates time-to-event outcomes.

Simulation of time-to-event outcomes using the piecewise constant hazard exponential function.

pw_exp_sim(hazard, n, maxtime = NULL, cutpoint = NULL)

Arguments

hazard
vector. The constant hazard rates for exponential failures.

n
scalar. The number of outcomes for simulation.

maxtime
scalar. maximum time before end of study.

cutpoint
vector. The change-point vector indicating time when the hazard rates change.

Value

a dataset with simulated follow-up time (time) and respective event indicator (1 = event, 0 = censoring)

Examples

pw_exp_sim(c(0.02, 0.01, 0.005), 100, 100, c(10, 20))

#>            time event
#> 1     0.3358204     1
#> 2   100.0000000     0
#> 3   100.0000000     0
#> 4   100.0000000     0
#> 5    58.1587937     1
#> 6   100.0000000     0
#> 7   100.0000000     0
#> 8    28.7605013     1
#> 9   100.0000000     0
#> 10    3.1730210     1
#> 11  100.0000000     0
#> 12  100.0000000     0
#> 13  100.0000000     0
#> 14    4.6201300     1
#> 15  100.0000000     0
#> 16   92.0929577     1
#> 17   81.7522866     1
#> 18  100.0000000     0
#> 19    6.3587477     1
#> 20   64.7032051     1
#> 21    5.6386188     1
#> 22   24.1326711     1
#> 23  100.0000000     0
#> 24  100.0000000     0
#> 25  100.0000000     0
#> 26    8.8093332     1
#> 27  100.0000000     0
#> 28   16.0583660     1
#> 29    6.6110411     1
#> 30   12.8375970     1
#> 31   40.5674197     1
#> 32  100.0000000     0
#> 33   74.4546706     1
#> 34   64.0549136     1
#> 35  100.0000000     0
#> 36  100.0000000     0
#> 37    7.8773326     1
#> 38    9.8452763     1
#> 39   15.8618438     1
#> 40  100.0000000     0
#> 41   44.9786483     1
#> 42  100.0000000     0
#> 43    1.6972963     1
#> 44  100.0000000     0
#> 45  100.0000000     0
#> 46   98.6312716     1
#> 47  100.0000000     0
#> 48   16.1829015     1
#> 49  100.0000000     0
#> 50    4.7703586     1
#> 51  100.0000000     0
#> 52    4.9033230     1
#> 53  100.0000000     0
#> 54  100.0000000     0
#> 55  100.0000000     0
#> 56   37.2051957     1
#> 57  100.0000000     0
#> 58  100.0000000     0
#> 59  100.0000000     0
#> 60  100.0000000     0
#> 61  100.0000000     0
#> 62  100.0000000     0
#> 63    0.4791993     1
#> 64   31.9065674     1
#> 65  100.0000000     0
#> 66   91.9310110     1
#> 67   19.5501783     1
#> 68  100.0000000     0
#> 69    6.2462562     1
#> 70  100.0000000     0
#> 71   35.5099139     1
#> 72  100.0000000     0
#> 73   10.6297712     1
#> 74   44.0961904     1
#> 75  100.0000000     0
#> 76   18.0079713     1
#> 77    0.6328134     1
#> 78  100.0000000     0
#> 79  100.0000000     0
#> 80  100.0000000     0
#> 81   75.2198568     1
#> 82    9.4428510     1
#> 83  100.0000000     0
#> 84  100.0000000     0
#> 85  100.0000000     0
#> 86  100.0000000     0
#> 87   75.1595106     1
#> 88  100.0000000     0
#> 89    7.5355268     1
#> 90   66.7296275     1
#> 91  100.0000000     0
#> 92   10.4348516     1
#> 93    1.6606189     1
#> 94  100.0000000     0
#> 95  100.0000000     0
#> 96   57.9291212     1
#> 97    7.6563546     1
#> 98  100.0000000     0
#> 99  100.0000000     0
#> 100 100.0000000     0

          pw_exp_sim(0.015, 100, 100)

#>            time event
#> 1    23.1718304     1
#> 2     6.1125170     1
#> 3    80.5559966     1
#> 4    70.3944252     1
#> 5    41.2911281     1
#> 6    36.1732871     1
#> 7    23.6901504     1
#> 8   100.0000000     0
#> 9    11.8972867     1
#> 10   14.5902656     1
#> 11   26.4705797     1
#> 12   65.5742928     1
#> 13   19.2992468     1
#> 14  100.0000000     0
#> 15    1.2827838     1
#> 16  100.0000000     0
#> 17   72.6195675     1
#> 18   18.8571301     1
#> 19   26.6261933     1
#> 20  100.0000000     0
#> 21  100.0000000     0
#> 22  100.0000000     0
#> 23   38.8624364     1
#> 24   98.6059587     1
#> 25   10.0630298     1
#> 26  100.0000000     0
#> 27   12.1402167     1
#> 28   71.0008699     1
#> 29  100.0000000     0
#> 30   75.2961172     1
#> 31  100.0000000     0
#> 32   28.9518838     1
#> 33   13.1593674     1
#> 34   23.5555757     1
#> 35  100.0000000     0
#> 36   16.5899198     1
#> 37   38.7826133     1
#> 38   36.4640988     1
#> 39  100.0000000     0
#> 40   67.9161935     1
#> 41    8.3719227     1
#> 42   82.3808017     1
#> 43   31.4527546     1
#> 44   10.0102170     1
#> 45    8.3518115     1
#> 46   78.0940170     1
#> 47    7.2693219     1
#> 48   94.3975162     1
#> 49   28.8943743     1
#> 50  100.0000000     0
#> 51   29.8785009     1
#> 52   70.4535330     1
#> 53   64.9183733     1
#> 54   29.2768081     1
#> 55   33.1620979     1
#> 56   14.1247329     1
#> 57   19.7509081     1
#> 58   17.1566667     1
#> 59    7.1991567     1
#> 60   52.1691425     1
#> 61   57.7019930     1
#> 62   56.5660990     1
#> 63   80.1003288     1
#> 64   52.1666182     1
#> 65   53.0551027     1
#> 66  100.0000000     0
#> 67   21.0833923     1
#> 68  100.0000000     0
#> 69   89.1836206     1
#> 70   18.0630561     1
#> 71  100.0000000     0
#> 72    0.8751236     1
#> 73  100.0000000     0
#> 74    0.3455069     1
#> 75    1.5471656     1
#> 76   87.7344128     1
#> 77   65.4086755     1
#> 78   54.3311908     1
#> 79   18.5830767     1
#> 80    7.8989181     1
#> 81   70.3387170     1
#> 82   40.3898500     1
#> 83  100.0000000     0
#> 84   34.7508421     1
#> 85   35.9312218     1
#> 86   36.0873358     1
#> 87   79.0104303     1
#> 88  100.0000000     0
#> 89   28.0470757     1
#> 90    4.8343500     1
#> 91   61.4629205     1
#> 92   39.5142228     1
#> 93   34.9817992     1
#> 94   18.1675164     1
#> 95   74.3773181     1
#> 96   63.3835391     1
#> 97    7.1960640     1
#> 98  100.0000000     0
#> 99   52.9397898     1
#> 100 100.0000000     0