Simulates time-to-event outcomes.

Simulation of time-to-event outcomes using the piecewise constant hazard exponential function.

pw_exp_sim(hazard, n, maxtime = NULL, cutpoint = NULL)

Arguments

hazard
vector. The constant hazard rates for exponential failures.

n
scalar. The number of outcomes for simulation.

maxtime
scalar. maximum time before end of study.

cutpoint
vector. The change-point vector indicating time when the hazard rates change.

Value

a dataset with simulated follow-up time (time) and respective event indicator (1 = event, 0 = censoring)

Examples

pw_exp_sim(c(0.02, 0.01, 0.005), 100, 100, c(10, 20))

#>           time event
#> 1   100.000000     0
#> 2   100.000000     0
#> 3    97.492922     1
#> 4   100.000000     0
#> 5   100.000000     0
#> 6   100.000000     0
#> 7     2.581297     1
#> 8   100.000000     0
#> 9   100.000000     0
#> 10  100.000000     0
#> 11  100.000000     0
#> 12    3.355043     1
#> 13  100.000000     0
#> 14   25.172474     1
#> 15   62.457992     1
#> 16    2.562327     1
#> 17   17.119260     1
#> 18   11.869983     1
#> 19   57.131759     1
#> 20  100.000000     0
#> 21  100.000000     0
#> 22  100.000000     0
#> 23  100.000000     0
#> 24   19.524019     1
#> 25  100.000000     0
#> 26  100.000000     0
#> 27   90.575245     1
#> 28   68.860190     1
#> 29  100.000000     0
#> 30  100.000000     0
#> 31    6.945050     1
#> 32    7.309035     1
#> 33   11.256203     1
#> 34   33.985024     1
#> 35  100.000000     0
#> 36  100.000000     0
#> 37    6.082908     1
#> 38  100.000000     0
#> 39   75.652219     1
#> 40  100.000000     0
#> 41  100.000000     0
#> 42  100.000000     0
#> 43   19.566333     1
#> 44   69.409398     1
#> 45   57.830281     1
#> 46  100.000000     0
#> 47  100.000000     0
#> 48    0.580311     1
#> 49   51.159356     1
#> 50  100.000000     0
#> 51  100.000000     0
#> 52   93.503282     1
#> 53    9.251223     1
#> 54   35.156444     1
#> 55   31.846300     1
#> 56  100.000000     0
#> 57  100.000000     0
#> 58    6.986779     1
#> 59  100.000000     0
#> 60   72.047840     1
#> 61  100.000000     0
#> 62    7.131641     1
#> 63    8.989694     1
#> 64   82.261494     1
#> 65  100.000000     0
#> 66  100.000000     0
#> 67   97.874384     1
#> 68  100.000000     0
#> 69  100.000000     0
#> 70  100.000000     0
#> 71  100.000000     0
#> 72  100.000000     0
#> 73  100.000000     0
#> 74  100.000000     0
#> 75  100.000000     0
#> 76   26.541675     1
#> 77   11.650866     1
#> 78  100.000000     0
#> 79  100.000000     0
#> 80  100.000000     0
#> 81  100.000000     0
#> 82   13.467446     1
#> 83   77.345496     1
#> 84  100.000000     0
#> 85  100.000000     0
#> 86  100.000000     0
#> 87   11.668381     1
#> 88    5.597290     1
#> 89   18.621180     1
#> 90  100.000000     0
#> 91  100.000000     0
#> 92   39.191164     1
#> 93  100.000000     0
#> 94    6.099694     1
#> 95  100.000000     0
#> 96  100.000000     0
#> 97   10.377092     1
#> 98  100.000000     0
#> 99   18.544695     1
#> 100  70.615190     1

          pw_exp_sim(0.015, 100, 100)

#>            time event
#> 1    64.7459108     1
#> 2    57.1504730     1
#> 3    17.6948933     1
#> 4    84.2827093     1
#> 5   100.0000000     0
#> 6    60.9072919     1
#> 7    67.6618055     1
#> 8    80.5435877     1
#> 9     1.2465748     1
#> 10  100.0000000     0
#> 11   92.2042398     1
#> 12   72.1843121     1
#> 13   23.3848469     1
#> 14   58.0367908     1
#> 15  100.0000000     0
#> 16   70.5007967     1
#> 17   19.6059973     1
#> 18  100.0000000     0
#> 19   43.9806052     1
#> 20    6.4471265     1
#> 21    9.0825133     1
#> 22  100.0000000     0
#> 23   20.6242915     1
#> 24   96.9725807     1
#> 25   73.8107602     1
#> 26   62.1244648     1
#> 27   38.8753932     1
#> 28  100.0000000     0
#> 29   38.3730992     1
#> 30   14.8769315     1
#> 31  100.0000000     0
#> 32   10.3853721     1
#> 33    3.5562250     1
#> 34    4.0961777     1
#> 35   44.1355500     1
#> 36  100.0000000     0
#> 37   13.2287859     1
#> 38   57.6540037     1
#> 39   24.0838187     1
#> 40  100.0000000     0
#> 41    7.7277147     1
#> 42  100.0000000     0
#> 43   72.4004755     1
#> 44   62.1155312     1
#> 45  100.0000000     0
#> 46   60.4838718     1
#> 47  100.0000000     0
#> 48   54.6778483     1
#> 49   14.4279663     1
#> 50   63.3583853     1
#> 51    6.0667342     1
#> 52  100.0000000     0
#> 53   18.7789851     1
#> 54   29.4698067     1
#> 55   20.8472203     1
#> 56   72.8540520     1
#> 57  100.0000000     0
#> 58    0.4535678     1
#> 59    1.9576839     1
#> 60   83.9609362     1
#> 61   66.9985893     1
#> 62   48.6658382     1
#> 63  100.0000000     0
#> 64  100.0000000     0
#> 65   44.1928954     1
#> 66   55.3969202     1
#> 67   50.7326336     1
#> 68  100.0000000     0
#> 69   27.7644734     1
#> 70   21.5358401     1
#> 71  100.0000000     0
#> 72   55.1046747     1
#> 73   37.4010893     1
#> 74   29.3139683     1
#> 75   24.7234125     1
#> 76   24.6948709     1
#> 77   12.8264342     1
#> 78   52.1021192     1
#> 79   46.1772036     1
#> 80    6.6124080     1
#> 81   27.1859786     1
#> 82  100.0000000     0
#> 83   24.8643341     1
#> 84  100.0000000     0
#> 85   66.2743347     1
#> 86  100.0000000     0
#> 87   74.8245362     1
#> 88   35.8180690     1
#> 89  100.0000000     0
#> 90  100.0000000     0
#> 91   50.0554632     1
#> 92   24.3290458     1
#> 93  100.0000000     0
#> 94   50.3113197     1
#> 95    0.2474643     1
#> 96  100.0000000     0
#> 97   14.1023564     1
#> 98    4.1211634     1
#> 99   18.0355721     1
#> 100   2.7949935     1