Simulates time-to-event outcomes.

Simulation of time-to-event outcomes using the piecewise constant hazard exponential function.

pw_exp_sim(hazard, n, maxtime = NULL, cutpoint = NULL)

Arguments

hazard
vector. The constant hazard rates for exponential failures.

n
scalar. The number of outcomes for simulation.

maxtime
scalar. maximum time before end of study.

cutpoint
vector. The change-point vector indicating time when the hazard rates change.

Value

a dataset with simulated follow-up time (time) and respective event indicator (1 = event, 0 = censoring)

Examples

pw_exp_sim(c(0.02, 0.01, 0.005), 100, 100, c(10, 20))

#>           time event
#> 1     1.007509     1
#> 2    96.034497     1
#> 3    10.994925     1
#> 4   100.000000     0
#> 5   100.000000     0
#> 6   100.000000     0
#> 7    24.957956     1
#> 8    17.460847     1
#> 9   100.000000     0
#> 10   98.876794     1
#> 11  100.000000     0
#> 12    9.248270     1
#> 13  100.000000     0
#> 14    9.801018     1
#> 15  100.000000     0
#> 16   98.456131     1
#> 17    8.244884     1
#> 18   60.323425     1
#> 19   10.299987     1
#> 20  100.000000     0
#> 21    6.410576     1
#> 22  100.000000     0
#> 23    7.409869     1
#> 24  100.000000     0
#> 25  100.000000     0
#> 26   29.027267     1
#> 27  100.000000     0
#> 28  100.000000     0
#> 29    8.631898     1
#> 30  100.000000     0
#> 31  100.000000     0
#> 32   60.188831     1
#> 33  100.000000     0
#> 34    4.947042     1
#> 35  100.000000     0
#> 36  100.000000     0
#> 37  100.000000     0
#> 38   96.751882     1
#> 39  100.000000     0
#> 40   30.068291     1
#> 41  100.000000     0
#> 42    3.739624     1
#> 43    2.888270     1
#> 44  100.000000     0
#> 45   93.927892     1
#> 46  100.000000     0
#> 47   11.078509     1
#> 48  100.000000     0
#> 49   31.308289     1
#> 50   11.115352     1
#> 51    9.820044     1
#> 52    4.509026     1
#> 53  100.000000     0
#> 54  100.000000     0
#> 55  100.000000     0
#> 56   52.503102     1
#> 57    7.970221     1
#> 58   57.591055     1
#> 59   35.712018     1
#> 60  100.000000     0
#> 61  100.000000     0
#> 62  100.000000     0
#> 63  100.000000     0
#> 64  100.000000     0
#> 65    7.252095     1
#> 66  100.000000     0
#> 67   56.712109     1
#> 68   91.504103     1
#> 69   10.154037     1
#> 70   25.101482     1
#> 71    7.762692     1
#> 72   73.134979     1
#> 73  100.000000     0
#> 74   79.464294     1
#> 75    4.818306     1
#> 76    9.442006     1
#> 77  100.000000     0
#> 78   16.267226     1
#> 79  100.000000     0
#> 80  100.000000     0
#> 81  100.000000     0
#> 82  100.000000     0
#> 83  100.000000     0
#> 84    4.916512     1
#> 85  100.000000     0
#> 86  100.000000     0
#> 87  100.000000     0
#> 88  100.000000     0
#> 89  100.000000     0
#> 90  100.000000     0
#> 91  100.000000     0
#> 92  100.000000     0
#> 93    7.934126     1
#> 94  100.000000     0
#> 95    6.250935     1
#> 96   33.515494     1
#> 97   43.519998     1
#> 98  100.000000     0
#> 99  100.000000     0
#> 100 100.000000     0

          pw_exp_sim(0.015, 100, 100)

#>             time event
#> 1    38.00374521     1
#> 2    41.59553172     1
#> 3    21.21047694     1
#> 4    59.85740548     1
#> 5     9.52570072     1
#> 6   100.00000000     0
#> 7   100.00000000     0
#> 8   100.00000000     0
#> 9     0.94045979     1
#> 10   56.79764748     1
#> 11   91.84289836     1
#> 12  100.00000000     0
#> 13    0.51341681     1
#> 14   19.33193611     1
#> 15    9.69551007     1
#> 16   76.84540755     1
#> 17  100.00000000     0
#> 18  100.00000000     0
#> 19   88.02833923     1
#> 20  100.00000000     0
#> 21   33.06404166     1
#> 22   31.38238195     1
#> 23  100.00000000     0
#> 24  100.00000000     0
#> 25  100.00000000     0
#> 26   64.29774253     1
#> 27  100.00000000     0
#> 28   98.72285034     1
#> 29   21.69178979     1
#> 30   53.37965519     1
#> 31   51.58026414     1
#> 32   26.98382450     1
#> 33   23.63055156     1
#> 34   24.69702354     1
#> 35    6.15180759     1
#> 36  100.00000000     0
#> 37    0.04630505     1
#> 38   46.66046283     1
#> 39   33.44988395     1
#> 40   93.42418901     1
#> 41   37.83224362     1
#> 42    4.79371514     1
#> 43   18.81132931     1
#> 44  100.00000000     0
#> 45   45.68697580     1
#> 46  100.00000000     0
#> 47   25.85088620     1
#> 48   80.80216633     1
#> 49  100.00000000     0
#> 50   74.42839661     1
#> 51   67.82639455     1
#> 52   10.60474099     1
#> 53  100.00000000     0
#> 54   46.76461598     1
#> 55   16.38549278     1
#> 56  100.00000000     0
#> 57   56.33727808     1
#> 58   58.84383252     1
#> 59   44.43220754     1
#> 60   46.30228660     1
#> 61  100.00000000     0
#> 62   85.31793244     1
#> 63  100.00000000     0
#> 64   24.42136944     1
#> 65   54.33889031     1
#> 66   56.36143169     1
#> 67    0.90474636     1
#> 68  100.00000000     0
#> 69   68.15126841     1
#> 70   26.79473233     1
#> 71  100.00000000     0
#> 72  100.00000000     0
#> 73   89.50525820     1
#> 74   67.59823455     1
#> 75   29.90274889     1
#> 76  100.00000000     0
#> 77    9.54107810     1
#> 78    6.02318507     1
#> 79  100.00000000     0
#> 80   31.35142596     1
#> 81  100.00000000     0
#> 82   22.97413752     1
#> 83   25.98292697     1
#> 84   45.56816764     1
#> 85   14.27629972     1
#> 86   14.92524061     1
#> 87   31.68151767     1
#> 88   72.14131777     1
#> 89   40.90074167     1
#> 90  100.00000000     0
#> 91   14.36177082     1
#> 92  100.00000000     0
#> 93   67.50134335     1
#> 94   83.11437598     1
#> 95   57.56724880     1
#> 96   10.13341596     1
#> 97   28.50080164     1
#> 98   45.46701675     1
#> 99   80.02921982     1
#> 100  31.50975474     1